refactor: create Runner class
