+++ src/main.ts
@@ -1,26 +1,13 @@
 #!/usr/bin/env node
 import yargs from 'yargs';
 import { hideBin } from 'yargs/helpers';
-import { globSync } from 'fast-glob';
-import { getProjectCompiledSources, Config } from './utils';
-import { Processor } from './processor';
+import { Config } from './types/config.t';
+import { Runner } from './runner';
 
 (async () => {
   const config: Config = getArguments();
-  const ignoredPaths = config.ignore.map((path) => globSync(path, { cwd: config.root })).flat();
-  const sourceUnits = await getProjectCompiledSources(config.root, config.contracts, ignoredPaths);
-  if (!sourceUnits.length) return console.error('No solidity files found in the specified directory');
-
-  const processor = new Processor(config);
-  const warnings = processor.processSources(sourceUnits);
-
-  warnings.forEach(({ location, messages }) => {
-    console.warn(location);
-    messages.forEach((message) => {
-      console.warn(`  ${message}`);
-    });
-    console.warn();
-  });
+  const runner: Runner = new Runner(config);
+  await runner.run();
 })().catch(console.error);
 
 function getArguments(): Config {
