+++ src/types/config.t.ts
@@ -1,0 +1,7 @@
+export interface Config {
+  root: string;
+  contracts: string;
+  enforceInheritdoc: boolean;
+  constructorNatspec: boolean;
+  ignore: string[];
+}
